@@ -1,11 +1,15 @@
 /** @type {import('tailwindcss').Config} */
-module.exports = {
-    content: [
-        "./src/**/*.{js,jsx,ts,tsx}",
-        "./public/index.html",
-    ],
-    theme: {
-        extend: {},
-    },
-    plugins: [],
-}
+export default {
+  content: [
+    "./index.html",
+    "./src/**/*.{js,jsx}"
+  ],
+  theme: {
+    extend: {
+      fontFamily: {
+        sans: ["system-ui", "ui-sans-serif", "SF Pro Text", "sans-serif"]
+      }
+    }
+  },
+  plugins: []
+};
